multi detector mode

+++ src/pmcxcl.cpp
@@ -1251,9 +1251,9 @@ py::dict pmcxcl_interface(const py::dict& user_cfg) {
 
         /** Initialize all buffers necessary to store the output variables */
         if (mcx_config.issave2pt == 1) {
-            int field_len =
-                static_cast<int>(mcx_config.dim.x) * static_cast<int>(mcx_config.dim.y) * static_cast<int>(mcx_config.dim.z) *
-                (int) ((mcx_config.tend - mcx_config.tstart) / mcx_config.tstep + 0.5) * mcx_config.srcnum;
+            size_t field_len =
+                static_cast<size_t>(mcx_config.dim.x) * static_cast<size_t>(mcx_config.dim.y) * static_cast<size_t>(mcx_config.dim.z) *
+                static_cast<size_t>((mcx_config.tend - mcx_config.tstart) / mcx_config.tstep + 0.5) * mcx_config.srcnum;
 
             if (mcx_config.replay.seed != nullptr && mcx_config.replaydet == -1) {
                 field_len *= mcx_config.detnum;
@@ -1387,7 +1387,7 @@ py::dict pmcxcl_interface(const py::dict& user_cfg) {
         }
 
         if (mcx_config.issave2pt) {
-            int field_len;
+            size_t field_len;
             field_dim[0] = mcx_config.srcnum * mcx_config.dim.x;
             field_dim[1] = mcx_config.dim.y;
             field_dim[2] = mcx_config.dim.z;
@@ -1418,19 +1418,19 @@ py::dict pmcxcl_interface(const py::dict& user_cfg) {
             auto dref_array = py::array_t<float, py::array::f_style>(array_dims);
 
             if (mcx_config.issaveref) {
-                int highdim = field_dim[3] * field_dim[4] * field_dim[5];
-                int voxellen = mcx_config.dim.x * mcx_config.dim.y * mcx_config.dim.z;
+                size_t highdim = field_dim[3] * field_dim[4] * field_dim[5];
+                size_t voxellen = static_cast<size_t>(mcx_config.dim.x) * mcx_config.dim.y * mcx_config.dim.z;
                 auto* dref = static_cast<float*>(dref_array.mutable_data());
                 memcpy(dref, mcx_config.exportfield, field_len * sizeof(float));
 
-                for (int voxelid = 0; voxelid < voxellen; voxelid++) {
+                for (size_t voxelid = 0; voxelid < voxellen; voxelid++) {
                     if (mcx_config.vol[voxelid]) {
-                        for (int gate = 0; gate < highdim; gate++)
+                        for (size_t gate = 0; gate < highdim; gate++)
                             for (unsigned int srcid = 0; srcid < mcx_config.srcnum; srcid++) {
                                 dref[(gate * voxellen + voxelid) * mcx_config.srcnum + srcid] = 0.f;
                             }
                     } else {
-                        for (int gate = 0; gate < highdim; gate++)
+                        for (size_t gate = 0; gate < highdim; gate++)
                             for (unsigned int srcid = 0; srcid < mcx_config.srcnum; srcid++) {
                                 dref[(gate * voxellen + voxelid) * mcx_config.srcnum + srcid] = -dref[(gate * voxellen + voxelid) * mcx_config.srcnum + srcid];
                                 mcx_config.exportfield[(gate * voxellen + voxelid) * mcx_config.srcnum + srcid] = 0.f;
